Add basic FOTA capable bootloader
This commit introduces a local lightweight bootloader that enables FOTA firmware updates.
It currently supports firmware updates over TCP via smoltcp IP stack, but stubs are in
place to support updates over CAN as well.

This is the new flash layout:

- first 8 bytes are used to store a sticky word for the bootloader flagging mechanism
- followed by the bootloader firmware at 0x2000_0008
- bootloader configurations are stored in sector 3, 0x0801_8000
- OxCC firmware stored in sector 4, 0x0802_0000

To enter the bootloader, reset the board with the user-button pressed or send the board
a bootloader reset CAN message.

A simple host-side `firmware-updater` Python script is provided as an example tool to
interract with the bootloader over a TCP connection.